fix: two regressions an audit found in tonight's own changes
An adversarial review of #96-#103 found two HIGH issues, both introduced by me
tonight. Both are fixed here, with the reproductions it gave.

--- F1: #102 refused values that are IN the allowed set -------------------

  atlasctl run <recipe> -o block_size=16
  error: `block_size` must be one of: 8, 16, 32, 64      <- 16 is in the list

  atlasctl run <recipe> -o enable_prefix_caching=1
  error: `enable_prefix_caching` expected true or false  <- worked before

`check_override` handed the CLI's YAML-typed scalar straight to `Bound::check`,
which accepts only `Str` for an Enum and only `Bool` for a toggle — because the
web surface sends what dropdowns and checkboxes produce. The CLI types `16` as
Int, so enforcing the bound rejected the natural spelling of a legal value, with
a message that contradicted itself.

Coerced now through `ScalarValue::render` and `is_truthy` — the SAME functions
the command renderer uses — so a bound can never refuse a value the renderer
would have accepted. Checked against the last commit before #102: block_size=16,
enable_prefix_caching=1 and =0 all render exactly as they did, and =0 still
omits the flag. Out-of-range values are still refused.

--- F2: #98 turned a wrong container-name guess into silent success -------

`container_of` guessed `atlas-{typed}`, but a container is named from the
RESOLVED recipe and a cluster launch appends `-rank{n}`. Once "no such
container" stopped being a failure, every missed guess became exit 0 while the
model was still serving:

  * `run X --rank 0` makes `atlas-X-rank0`, and run.rs PRINTS `atlasctl stop X`
    as the way to stop it — which then reported "not running", exit 0.
  * `stop @registry/X` guessed `atlas-@registry/X`.

A reboot script doing per-recipe stops on a two-node fleet was told the box was
idle. Before #98 this failed loudly; #98 made it quiet, which is worse.

`stop <recipe>` now finds containers by `label=io.atlasctl.recipe=<resolved>`,
which the launch itself writes. No name arithmetic, so ranks and scoped refs
both work, and the label survives a registry being removed. This is the house
pattern already asserted by `managed_containers_are_found_by_label_not_by_name_guessing`
— which `stop` was quietly violating.

`stop` splits into `stop_all_with` and `stop_recipe_with` so there is no
unreachable branch left over from the old guess. Two tests added for exactly
the reported cases (a rank container, a scoped ref), plus the not-running test
rewritten: it was passing for the wrong reason after the change, because its
pushed result was being consumed by the label query rather than by `docker stop`.

--- and the cap --------------------------------------------------------

lifecycle.rs landed at exactly 500 lines, which passes `-gt 500` but leaves no
room. Tests moved to `lifecycle/tests.rs` verbatim, the house pattern: 259 and
246 lines. The inner module is renamed `naming` because a `tests` inside
`tests.rs` shadows its parent.

Workspace green (11 suites, 119 in atlasctl), clippy and fmt clean, goldens
unchanged, no file over the cap.
